+++ tests/support/class-auth-service-test-case.php
@@ -30,7 +30,8 @@ abstract class AuthServiceTestCase extends TestCase {
 			$GLOBALS['alynt_ag_test_existing_emails'],
 			$GLOBALS['alynt_ag_test_retrieve_password_result'],
 			$GLOBALS['alynt_ag_test_signon_roles'],
-			$GLOBALS['alynt_ag_test_nonce_valid']
+			$GLOBALS['alynt_ag_test_nonce_valid'],
+			$GLOBALS['alynt_ag_test_nocache_headers']
 		);
 		$_SERVER['REMOTE_ADDR'] = '203.0.113.30';
 		$_SERVER['REQUEST_METHOD'] = 'GET';